new splits are open to the current directory
If you split the current window, the new pane will start off in the same directory.
If you make a new window it will start off from whatever directory tmux was launched in, that behavior isn't affected. This makes it easier for me to setup a workspace in a particular directory.