Don't recursively map all the time
If users remap keys, the behaviour of the things set up in vimrc may change.

An example of this is the common remapping to switch `;` and `:`, which will
cause all of the keyboard shortcuts involving `:` to fail. Switching the *map
statements to their *noremap counterparts fixes this problem, by using the
original meanings.