add vim-lastplace
TLDR: Intelligently reopen files where you left off.

This is a feature I really enjoy having and this script has been very carefully crafted to be better than most of the snippets floating around the internet. For example, it won't be confused by commit messages and it will keep as much of the file on screen as possible if you were working at the end of a file.

I may be biased because I did create this plugin. Submitting this pull request to start a discussion about including it. 

Thanks!