Allow for `tmux` versions ending in "a"
If the tmux version ends in "a", as it does today with version 3.3a,
your `~/.tmux.conf` will break and you will be sad.

I'm removing the check for version 2.4 because it shipped a while ago
and allows me to remove the dependency on `bc`.

Here's where I figured this all out:
https://github.com/tmux/tmux/issues/3219#issuecomment-1151776848